+++ nerfbaselines/methods/gaussian_splatting_spec.py
@@ -48,7 +48,7 @@ GaussianSplattingSpec: MethodSpec = {
 cd gaussian-splatting
 git checkout 2eee0e26d2d5fd00ec462df47752223952f6bf4e
 
-conda install -y mkl==2023.1.0 pytorch==2.0.1 torchvision==0.15.2 pytorch-cuda=11.7 -c pytorch -c nvidia
+conda install -y mkl==2023.1.0 pytorch==2.0.1 torchvision==0.15.2 pytorch-cuda=11.7 'numpy<2.0.0' -c pytorch -c nvidia
 conda install -y cudatoolkit-dev=11.7 gcc_linux-64=11 gxx_linux-64=11 make=4.3 cmake=3.28.3 -c conda-forge
 pip install -U pip 'setuptools<70.0.0'
 pip install plyfile==0.8.1 tqdm submodules/diff-gaussian-rasterization submodules/simple-knn
