+++ nerfbaselines/methods/wild_gaussians_spec.py
@@ -19,7 +19,7 @@ if [ "$NB_DOCKER_BUILD" != "1" ]; then
 conda install -y gcc_linux-64=11 gxx_linux-64=11 make=4.3 cmake=3.28.3 -c conda-forge
 fi
 pip install --upgrade pip
-pip install -r requirements.txt
+pip install 'numpy<2.0.0' -r requirements.txt
 LIBRARY_PATH="$CONDA_PREFIX/lib/stubs" pip install -e ./submodules/diff-gaussian-rasterization ./submodules/simple-knn
 pip install -e .
 pip install opencv-python-headless
